fix: split coordinator, fix stale captures, update AGENTS.md scroll guidance
Splits MessageListScrollCoordinator into 3 extension files per AGENTS.md
~500-600 line target. Fixes stale conversationId capture in pin coordinator
closures by reading live coordinator state. Removes unused isResizeActive
parameter. Updates AGENTS.md scroll guidance to reflect new architecture.
